requirements.txt
protobuf>=4.25.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
google-adk>=0.1.0
gunicorn>=21.2.0
flask>=3.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
pydantic>=2.5.0
typing-extensions>=4.8.0
